+++ Common/UI/CraftableUISystem.cs
@@ -44,6 +44,7 @@ namespace SteroidGuide.Common.UI
         {
             ItemScanner.ClearSyncState();
             CraftableState?.CancelPendingAnalysis();
+            MagicStorageBridge.ResetGate();
         }
 
         public void ShowUI(int npcIndex = -1)